requirements.txt
image-quality==1.2.7
imageio==2.37.0
lazy_loader==0.4
libsvm==3.23.0.4
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==12.0.0
scikit-image==0.25.2
scipy==1.16.3
tifffile==2025.10.16
